phase 7: show-ready playback — fade engine, stage capture, gobo DMX, cue editing, MIDI/OSC triggers, unified .slx show file
Phase 7.1 — Fade Engine
- PlayheadState::Fading with CueValues::lerp() for all attributes
- Per-fixture interpolation in cue_to_dmx() with auto-transition to Idle
- Fade progress UI bar in cue panel

Phase 7.2 — Stage Capture
- CaptureMode (Programmer/Stage) toggle in cue panel
- RecordStageCueEvent observer inverse-projects merged DMX output per fixture
- Inverse DMX projection for dimmer, pan/tilt (8/16-bit), RGB, zoom

Phase 7.3 — Attribute Coverage
- DmxChannelMap extended with gobo, gobo_rotation, color_wheel
- CueValues got gobo_index/gobo_spin; forward/inverse projection updated
- GDTF parser maps Gobo, GoboRot, ColorWheel attributes

Phase 7.4 — Cue Editing
- LoadCueIntoProgrammerEvent: click cue row to load for editing
- UpdateCueEvent: UPDATE button overwrites cue snapshot
- Inline label edit + fade in/out DragValues (0–60s)

Phase 7.5 — MIDI/OSC Triggers
- MidiConfig note_go/note_back; midi_receive() Note On handling
- osc_receive() /cue/go and /cue/back handlers
- IO panel UI for note selectors with enable checkboxes

Phase 7.6 — Unified Show File & UI Stubs
- ShowFile struct bundles patch, cues, venue into .slx.json
- SaveShowEvent / LoadShowEvent with observers and auto-load on startup
- Ctrl+S / Ctrl+O keyboard shortcuts with rfd file dialog
- Status bar wired: FPS, CPU%, universe usage, venue name from real data
- Removed dead RuntimeStats resource, all persistence goes through .slx
- Backward-compat: auto-migrates legacy show.json

+++ crates/stagelx-render/src/scene.rs
@@ -58,7 +58,7 @@ pub fn setup_scene(
 ) {
     let (w, h, sf) = windows
         .single()
-        .map(|win| (win.physical_width(), win.physical_height(), win.scale_factor() as f32))
+        .map(|win| (win.physical_width(), win.physical_height(), win.scale_factor()))
         .unwrap_or((1440, 900, 2.0));
 
     let (foh_vp, top_vp, side_vp) = compute_viewports(w, h, sf);
@@ -145,6 +145,7 @@ pub fn setup_scene(
     });
 }
 
+#[allow(clippy::type_complexity)]
 pub fn update_viewports_on_resize(
     mut resize_events: MessageReader<WindowResized>,
     windows: Query<&Window>,
@@ -168,7 +169,7 @@ pub fn update_viewports_on_resize(
     if w == 0 || h == 0 {
         return;
     }
-    let sf = window.scale_factor() as f32;
+    let sf = window.scale_factor();
     let (foh_vp, top_vp, side_vp) = compute_viewports(w, h, sf);
 
     for (mut cam, foh, top, side) in &mut cameras {
